Use persist.js (and all its backends) properly.
It seems not all browsers were defaulting to
persist.js' localStorage backend.
This re-implements large chunks of the plugin
to use the persist.js API properly, and thus
make it backend agnostic.

One downside to this is that the API is async.
This makes it tricky to guarantee population of
the helper variables from *sm_update ahead of CS
code referencing them. I've included an updated
README and example to try and explain how to
circumvent this, but it would benefit from
further investigation/improvement.

This patch also sneaks in a 'toggle' option
for *sm_menu.